requirements.txt
# Core environment
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4

# LangChain core + community integrations
langchain==0.3.3
langchain-core==0.3.15
langchain-community==0.3.2
langchain-openai==0.2.3
langchainhub==0.1.21

# Vector database
chromadb==0.5.4
faiss-cpu

# Document loaders and text processing
PyPDF2==3.0.1
pypdf==4.3.1
unstructured==0.15.8
pdfminer.six==20221105
tiktoken==0.5.1
python-docx==1.2.0

# Model embeddings and NLP
sentence-transformers==3.1.1
huggingface-hub==0.25.2

# Web frameworks (for UI)
streamlit==1.38.0
flask==3.0.3

# API integration & HTTP calls
requests==2.32.3
httpx==0.27.2

# Job data handling
beautifulsoup4==4.12.3
lxml==5.3.0

# Utils for agents & orchestration
tqdm==4.66.4
fastapi==0.115.2
uvicorn==0.30.6

# Optional visualization and debugging
# matplotlib==3.9.2
plotly==5.24.1
rich==13.8.1

# For notebook / development use
ipykernel==6.29.5


azure-identity==1.15.0